llm_query_retriever/requirements.txt
streamlit
PyMuPDF
python-docx
sentence-transformers
faiss-cpu
chromadb
google-generativeai
unstructured 
python-dotenv 